requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Azure services
azure-communication-email==1.0.0
azure-storage-blob==12.22.0
azure-identity==1.18.0

# Auth
msal==1.31.0
python-jose[cryptography]==3.3.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
apscheduler==3.10.4
pydantic[email]==2.9.2
pydantic-settings==2.5.2
